Rename test files (#17278)
