+++ mongo/tests/common.py
@@ -282,3 +282,30 @@ TLS_METADATA = {
         f'{TLS_CERTS_FOLDER}:/certs',
     ],
 }
+
+METRIC_VAL_CHECKS = {
+    'mongodb.asserts.msgps': lambda x: x >= 0,
+    'mongodb.fsynclocked': lambda x: x >= 0,
+    'mongodb.globallock.activeclients.readers': lambda x: x >= 0,
+    'mongodb.metrics.cursor.open.notimeout': lambda x: x >= 0,
+    'mongodb.metrics.document.deletedps': lambda x: x >= 0,
+    'mongodb.metrics.getlasterror.wtime.numps': lambda x: x >= 0,
+    'mongodb.metrics.repl.apply.batches.numps': lambda x: x >= 0,
+    'mongodb.metrics.ttl.deleteddocumentsps': lambda x: x >= 0,
+    'mongodb.network.bytesinps': lambda x: x >= 1,
+    'mongodb.network.numrequestsps': lambda x: x >= 1,
+    'mongodb.opcounters.commandps': lambda x: x >= 1,
+    'mongodb.opcountersrepl.commandps': lambda x: x >= 0,
+    'mongodb.oplog.logsizemb': lambda x: x >= 1,
+    'mongodb.oplog.timediff': lambda x: x >= 1,
+    'mongodb.oplog.usedsizemb': lambda x: x >= 0,
+    'mongodb.replset.health': lambda x: x >= 1,
+    'mongodb.replset.state': lambda x: x >= 1,
+    'mongodb.stats.avgobjsize': lambda x: x >= 0,
+    'mongodb.stats.storagesize': lambda x: x >= 0,
+    'mongodb.connections.current': lambda x: x >= 1,
+    'mongodb.connections.available': lambda x: x >= 1,
+    'mongodb.uptime': lambda x: x >= 0,
+    'mongodb.mem.resident': lambda x: x > 0,
+    'mongodb.mem.virtual': lambda x: x > 0,
+}
