+++ mongo/tests/test_integration_standalone.py
@@ -1,43 +1,18 @@
-# (C) Datadog, Inc. 2018-present
+# (C) Datadog, Inc. 2024-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import logging
 
-import mock
 import pytest
 
 from datadog_checks.dev.utils import get_metadata_metrics
 from datadog_checks.mongo import MongoDb
-from datadog_checks.mongo.common import ReplicaSetDeployment
 
 from . import common
+from .common import METRIC_VAL_CHECKS
 
-METRIC_VAL_CHECKS = {
-    'mongodb.asserts.msgps': lambda x: x >= 0,
-    'mongodb.fsynclocked': lambda x: x >= 0,
-    'mongodb.globallock.activeclients.readers': lambda x: x >= 0,
-    'mongodb.metrics.cursor.open.notimeout': lambda x: x >= 0,
-    'mongodb.metrics.document.deletedps': lambda x: x >= 0,
-    'mongodb.metrics.getlasterror.wtime.numps': lambda x: x >= 0,
-    'mongodb.metrics.repl.apply.batches.numps': lambda x: x >= 0,
-    'mongodb.metrics.ttl.deleteddocumentsps': lambda x: x >= 0,
-    'mongodb.network.bytesinps': lambda x: x >= 1,
-    'mongodb.network.numrequestsps': lambda x: x >= 1,
-    'mongodb.opcounters.commandps': lambda x: x >= 1,
-    'mongodb.opcountersrepl.commandps': lambda x: x >= 0,
-    'mongodb.oplog.logsizemb': lambda x: x >= 1,
-    'mongodb.oplog.timediff': lambda x: x >= 1,
-    'mongodb.oplog.usedsizemb': lambda x: x >= 0,
-    'mongodb.replset.health': lambda x: x >= 1,
-    'mongodb.replset.state': lambda x: x >= 1,
-    'mongodb.stats.avgobjsize': lambda x: x >= 0,
-    'mongodb.stats.storagesize': lambda x: x >= 0,
-    'mongodb.connections.current': lambda x: x >= 1,
-    'mongodb.connections.available': lambda x: x >= 1,
-    'mongodb.uptime': lambda x: x >= 0,
-    'mongodb.mem.resident': lambda x: x > 0,
-    'mongodb.mem.virtual': lambda x: x > 0,
-}
+pytestmark = [pytest.mark.usefixtures('dd_environment'), pytest.mark.integration, common.standalone]
+
 
 METRIC_VAL_CHECKS_OLD = {
     'mongodb.connections.current': lambda x: x >= 1,
@@ -48,10 +23,6 @@ METRIC_VAL_CHECKS_OLD = {
 }
 
 
-pytestmark = [pytest.mark.usefixtures('dd_environment'), pytest.mark.integration]
-
-
-@common.standalone
 @pytest.mark.parametrize(
     'instance_authdb',
     [
@@ -74,7 +45,6 @@ def test_mongo_authdb(aggregator, check, instance_authdb, dd_run_check):
     aggregator.assert_metrics_using_metadata(get_metadata_metrics(), check_submission_type=True)
 
 
-@common.standalone
 @pytest.mark.parametrize(
     'instance_user',
     [pytest.param(common.INSTANCE_USER, id='standard'), pytest.param(common.INSTANCE_USER_LEGACY_CONFIG, id='legacy')],
@@ -96,40 +66,6 @@ def test_mongo_db_test(aggregator, check, instance_user, dd_run_check):
     aggregator.assert_metrics_using_metadata(get_metadata_metrics(), check_submission_type=True)
 
 
-@common.shard
-def test_mongo_arbiter(aggregator, check, instance_arbiter, dd_run_check):
-    check = check(instance_arbiter)
-    dd_run_check(check)
-
-    tags = [f'host:{common.HOST}', f'port:{common.PORT_ARBITER}', 'db:admin']
-    aggregator.assert_service_check('mongodb.can_connect', status=MongoDb.OK, tags=tags)
-
-    metric_names = aggregator.metric_names
-    assert metric_names
-
-    for metric_name in metric_names:
-        if metric_name in METRIC_VAL_CHECKS:
-            metric = aggregator.metrics(metric_name)[0]
-            assert METRIC_VAL_CHECKS[metric_name](metric.value)
-    aggregator.assert_metrics_using_metadata(get_metadata_metrics(), check_submission_type=True)
-
-    expected_metrics = {
-        'mongodb.replset.health': 1.0,
-        'mongodb.replset.votefraction': None,
-        'mongodb.replset.votes': 1,
-        'mongodb.replset.state': 7,
-    }
-    expected_tags = [
-        'server:mongodb://localhost:27020/',
-        'replset_name:shard01',
-        'replset_state:arbiter',
-        'sharding_cluster_role:shardsvr',
-    ]
-    for metric, value in expected_metrics.items():
-        aggregator.assert_metric(metric, value, expected_tags, count=1)
-
-
-@common.standalone
 def test_mongo_old_config(aggregator, check, instance, dd_run_check):
     check = check(instance)
     dd_run_check(check)
@@ -144,7 +80,6 @@ def test_mongo_old_config(aggregator, check, instance, dd_run_check):
     aggregator.assert_metrics_using_metadata(get_metadata_metrics(), check_submission_type=True)
 
 
-@common.standalone
 def test_mongo_1valid_and_1invalid_custom_queries(
     aggregator, check, instance_1valid_and_1invalid_custom_queries, dd_run_check
 ):
@@ -157,7 +92,6 @@ def test_mongo_1valid_and_1invalid_custom_queries(
     aggregator.assert_metric("dd.custom.mongo.query_a.amount", count=0)
 
 
-@common.standalone
 def test_mongo_custom_queries(aggregator, check, instance_custom_queries, dd_run_check):
     # Run the check against our running server
     check = check(instance_custom_queries)
@@ -195,7 +129,6 @@ def test_mongo_custom_queries(aggregator, check, instance_custom_queries, dd_run
     aggregator.assert_metric('dd.mongodb.custom.queries_slower_than_60sec.secs_running', metric_type=aggregator.GAUGE)
 
 
-@common.standalone
 def test_mongo_custom_query_with_empty_result_set(aggregator, check, instance_user, caplog, dd_run_check):
     instance_user['custom_queries'] = [
         {
@@ -221,7 +154,6 @@ def test_mongo_custom_query_with_empty_result_set(aggregator, check, instance_us
     aggregator.assert_metric('dd.custom.mongo.query_a.amount', count=0)
 
 
-@common.standalone
 @pytest.mark.parametrize(
     'instance',
     [
@@ -243,7 +175,6 @@ def test_mongo_custom_query_with_date(aggregator, check, instance, dd_run_check)
     aggregator.assert_metric_has_tag("dd.custom.mongo.aggregate.total", 'tag2:val2', count=2)
 
 
-@common.standalone
 @pytest.mark.parametrize(
     'instance',
     [
@@ -260,35 +191,6 @@ def test_mongo_custom_query_with_string_list(aggregator, check, instance, dd_run
     aggregator.assert_metric_has_tag("dd.custom.mongo.string.result", 'collection:orders', count=4)
 
 
-@common.shard
-def test_mongo_replset(instance_shard, aggregator, check, dd_run_check):
-    mongo_check = check(instance_shard)
-    dd_run_check(mongo_check)
-
-    replset_metrics = [
-        'mongodb.replset.health',
-        'mongodb.replset.replicationlag',
-        'mongodb.replset.state',
-        'mongodb.replset.votefraction',
-        'mongodb.replset.votes',
-    ]
-    replset_common_tags = [
-        "replset_name:shard01",
-        "server:mongodb://localhost:27018/",
-        "sharding_cluster_role:shardsvr",
-    ]
-    for metric in replset_metrics:
-        aggregator.assert_metric(metric, tags=replset_common_tags + ['replset_state:primary'])
-    aggregator.assert_metric(
-        'mongodb.replset.optime_lag', tags=replset_common_tags + ['replset_state:primary', 'member:shard01a:27018']
-    )
-    aggregator.assert_metric(
-        'mongodb.replset.optime_lag', tags=replset_common_tags + ['replset_state:secondary', 'member:shard01b:27019']
-    )
-    aggregator.assert_metrics_using_metadata(get_metadata_metrics(), check_submission_type=True)
-
-
-@common.standalone
 def test_metadata(check, instance, datadog_agent):
     check = check(instance)
     check.check_id = 'test:123'
@@ -298,15 +200,3 @@ def test_metadata(check, instance, datadog_agent):
     check.check(instance)
     datadog_agent.assert_metadata('test:123', version_metadata)
     datadog_agent.assert_metadata_count(len(version_metadata) + 2)
-
-
-@common.shard
-def test_refresh_role(instance_shard, aggregator, check, dd_run_check):
-    mongo_check = check(instance_shard)
-    dd_run_check(mongo_check)
-    with mock.patch('datadog_checks.mongo.api.MongoApi._get_rs_deployment_from_status_payload') as get_deployment:
-        mock_deployment_type = ReplicaSetDeployment("sharding01", 9, cluster_role="TEST")
-        get_deployment.return_value = mock_deployment_type
-        dd_run_check(mongo_check)
-        assert get_deployment.call_count == 1
-        assert mongo_check.api_client.deployment_type.cluster_role == "TEST"
