+++ mongo/tests/test_unit_diagnose.py
@@ -1,4 +1,4 @@
-# (C) Datadog, Inc. 2023-present
+# (C) Datadog, Inc. 2024-present
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import json
